requirements.txt
streamlit
tensorflow
pillow 